feat: implement RootStoreProvider for improved state management in FileUploader
