+++ packages/pluggableWidgets/file-uploader-web/src/components/FileEntry.tsx
@@ -37,7 +37,6 @@ export const FileEntryContainer = observer(({ store, actions }: FileEntryContain
             mimeType={store.mimeType}
             fileStatus={store.fileStatus}
             errorMessage={store.errorDescription}
-            maxTotalFiles={store.maxTotalFiles}
             defaultAction={defaultListAction && store.canExecute(defaultListAction) ? onDefaultAction : undefined}
             actions={<ActionsBar actions={actions} store={store} />}
         />
@@ -52,7 +51,6 @@ interface FileEntryProps {
 
     fileStatus: FileStatus;
     errorMessage?: string;
-    maxTotalFiles: number;
 
     defaultAction?: () => void;
 
@@ -117,7 +115,7 @@ function FileEntry(props: FileEntryProps): ReactElement {
                 <ProgressBar visible={props.fileStatus === "uploading"} indeterminate />
             </div>
             <div className={"entry-upload-info"}>
-                <UploadInfo status={props.fileStatus} error={props.errorMessage} maxTotalFiles={props.maxTotalFiles} />
+                <UploadInfo status={props.fileStatus} error={props.errorMessage} />
             </div>
         </div>
     );
