--- packages/pluggableWidgets/file-uploader-web/src/utils/useRootStore.ts
@@ -1,21 +1,0 @@
-import { useEffect, useState } from "react";
-import { useTranslationsStore } from "./useTranslationsStore";
-import { FileUploaderContainerProps } from "../../typings/FileUploaderProps";
-import { FileUploaderStore } from "../stores/FileUploaderStore";
-
-export function useRootStore(props: FileUploaderContainerProps): FileUploaderStore {
-    const translations = useTranslationsStore();
-    const [rootStore] = useState(() => {
-        return new FileUploaderStore(props, translations);
-    });
-
-    useEffect(() => {
-        rootStore.updateProps(props);
-    }, [rootStore, props]);
-
-    useEffect(() => {
-        return () => rootStore.dispose();
-    }, [rootStore]);
-
-    return rootStore;
-}
